Utils: move package utils/moreatomic to internal/moreatomic
