+++ internal/moreatomic/bool.go
