+++ internal/moreatomic/string.go
