+++ utils/wsutil/op.go
@@ -5,9 +5,10 @@ import (
 	"fmt"
 	"sync"
 
-	"github.com/diamondburned/arikawa/utils/json"
-	"github.com/diamondburned/arikawa/utils/moreatomic"
 	"github.com/pkg/errors"
+
+	"github.com/diamondburned/arikawa/internal/moreatomic"
+	"github.com/diamondburned/arikawa/utils/json"
 )
 
 var ErrEmptyPayload = errors.New("empty payload")
